@@ -1,6 +1,6 @@
 {
   "name": "axios",
-  "version": "1.6.7",
+  "version": "1.7.2",
   "description": "Promise based HTTP client for the browser and node.js",
   "main": "index.js",
   "exports": {
@@ -80,41 +80,41 @@
   },
   "homepage": "https://axios-http.com",
   "devDependencies": {
-    "@babel/core": "^7.18.2",
-    "@babel/preset-env": "^7.18.2",
-    "@commitlint/cli": "^17.3.0",
-    "@commitlint/config-conventional": "^17.3.0",
+    "@babel/core": "^7.23.9",
+    "@babel/preset-env": "^7.23.9",
+    "@commitlint/cli": "^17.8.1",
+    "@commitlint/config-conventional": "^17.8.1",
     "@release-it/conventional-changelog": "^5.1.1",
     "@rollup/plugin-babel": "^5.3.1",
     "@rollup/plugin-commonjs": "^15.1.0",
     "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-multi-entry": "^4.0.0",
+    "@rollup/plugin-multi-entry": "^4.1.0",
     "@rollup/plugin-node-resolve": "^9.0.0",
-    "abortcontroller-polyfill": "^1.7.3",
+    "abortcontroller-polyfill": "^1.7.5",
     "auto-changelog": "^2.4.0",
-    "body-parser": "^1.20.0",
-    "chalk": "^5.2.0",
+    "body-parser": "^1.20.2",
+    "chalk": "^5.3.0",
     "coveralls": "^3.1.1",
     "cross-env": "^7.0.3",
     "dev-null": "^0.1.1",
     "dtslint": "^4.2.1",
     "es6-promise": "^4.2.8",
-    "eslint": "^8.17.0",
-    "express": "^4.18.1",
-    "formdata-node": "^5.0.0",
-    "formidable": "^2.0.1",
+    "eslint": "^8.56.0",
+    "express": "^4.18.2",
+    "formdata-node": "^5.0.1",
+    "formidable": "^2.1.2",
     "fs-extra": "^10.1.0",
     "get-stream": "^3.0.0",
     "gulp": "^4.0.2",
     "gzip-size": "^7.0.0",
-    "handlebars": "^4.7.7",
-    "husky": "^8.0.2",
+    "handlebars": "^4.7.8",
+    "husky": "^8.0.3",
     "istanbul-instrumenter-loader": "^3.0.1",
-    "jasmine-core": "^2.4.1",
+    "jasmine-core": "^2.99.1",
     "karma": "^6.3.17",
-    "karma-chrome-launcher": "^3.1.1",
+    "karma-chrome-launcher": "^3.2.0",
     "karma-firefox-launcher": "^2.1.2",
-    "karma-jasmine": "^1.1.1",
+    "karma-jasmine": "^1.1.2",
     "karma-jasmine-ajax": "^0.1.13",
     "karma-rollup-preprocessor": "^7.0.8",
     "karma-safari-launcher": "^1.0.0",
@@ -122,12 +122,12 @@
     "karma-sinon": "^1.0.5",
     "karma-sourcemap-loader": "^0.3.8",
     "memoizee": "^0.4.15",
-    "minimist": "^1.2.7",
-    "mocha": "^10.0.0",
+    "minimist": "^1.2.8",
+    "mocha": "^10.3.0",
     "multer": "^1.4.4",
-    "pretty-bytes": "^6.0.0",
-    "release-it": "^15.5.1",
-    "rollup": "^2.67.0",
+    "pretty-bytes": "^6.1.1",
+    "release-it": "^15.11.0",
+    "rollup": "^2.79.1",
     "rollup-plugin-auto-external": "^2.0.0",
     "rollup-plugin-bundle-size": "^1.0.3",
     "rollup-plugin-terser": "^7.0.2",
@@ -135,7 +135,8 @@
     "stream-throttle": "^0.1.3",
     "string-replace-async": "^3.0.2",
     "terser-webpack-plugin": "^4.2.3",
-    "typescript": "^4.8.4"
+    "typescript": "^4.9.5",
+    "@rollup/plugin-alias": "^5.1.0"
   },
   "browser": {
     "./lib/adapters/http.js": "./lib/helpers/null.js",
@@ -146,7 +147,7 @@
   "unpkg": "dist/axios.min.js",
   "typings": "./index.d.ts",
   "dependencies": {
-    "follow-redirects": "^1.15.4",
+    "follow-redirects": "^1.15.6",
     "form-data": "^4.0.0",
     "proxy-from-env": "^1.1.0"
   },
@@ -159,8 +160,8 @@
   "contributors": [
     "Matt Zabriskie (https://github.com/mzabriskie)",
     "Nick Uraltsev (https://github.com/nickuraltsev)",
-    "Jay (https://github.com/jasonsaayman)",
     "Dmitriy Mozgovoy (https://github.com/DigitalBrainJS)",
+    "Jay (https://github.com/jasonsaayman)",
     "Emily Morehouse (https://github.com/emilyemorehouse)",
     "Rubén Norte (https://github.com/rubennorte)",
     "Justin Beckwith (https://github.com/JustinBeckwith)",
