@@ -1,6 +1,7 @@
 {
   "name": "magic-string",
-  "version": "0.30.8",
+  "version": "0.30.10",
+  "packageManager": "pnpm@9.0.1",
   "description": "Modify strings, generate sourcemaps",
   "keywords": [
     "string",
@@ -29,10 +30,11 @@
     "README.md"
   ],
   "scripts": {
-    "build": "rollup -c && cp ./src/index.d.ts ./dist/magic-string.es.d.mts && cp ./src/index.d.ts ./dist/magic-string.cjs.d.ts",
+    "build": "rollup -c",
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s",
     "format": "prettier --single-quote --print-width 100 --use-tabs --write src/*.js src/**/*.js",
     "lint": "eslint src test && publint",
+    "lint:fix": "eslint src test --fix",
     "prepare": "npm run build",
     "prepublishOnly": "npm run lint && rm -rf dist && npm test",
     "release": "bumpp -x \"npm run changelog\" --all --commit --tag --push && npm publish",
@@ -45,18 +47,15 @@
     "@rollup/plugin-node-resolve": "^15.2.3",
     "@rollup/plugin-replace": "^5.0.5",
     "benchmark": "^2.1.4",
-    "bumpp": "^9.3.0",
+    "bumpp": "^9.4.0",
     "conventional-changelog-cli": "^3.0.0",
-    "eslint": "^8.56.0",
-    "mocha": "^10.2.0",
-    "prettier": "^3.2.4",
+    "eslint": "^8.57.0",
+    "mocha": "^10.4.0",
+    "prettier": "^3.2.5",
     "publint": "^0.2.7",
-    "rollup": "^3.28.0",
-    "source-map-js": "^1.0.2",
+    "rollup": "^3.29.4",
+    "source-map-js": "^1.2.0",
     "source-map-support": "^0.5.21"
-  },
-  "engines": {
-    "node": ">=12"
   },
   "dependencies": {
     "@jridgewell/sourcemap-codec": "^1.4.15"
